+++ ECChildMS/src/Components/Admin/Note.jsx
@@ -21,16 +21,18 @@ const Note = () => {
     }, [])
 
     const handleDelete = (id) => {
-        axios.delete('http://localhost:3000/auth/delete_note/' + id)
-            .then(result => {
-                if (result.data.Status) {
-                    setUserInput(userInput.filter(e => e.id !== id))
-                    window.location.reload()
-                } else {
-                    alert(result.data.Error)
-                }
-            })
-            .catch(err => console.log(err))
+        if (window.confirm("ALERT! Are you sure you want to delete this note?")) {
+            axios.delete('http://localhost:3000/auth/delete_note/' + id)
+                .then(result => {
+                    if (result.data.Status) {
+                        setUserInput(userInput.filter(e => e.id !== id))
+                        window.location.reload()
+                    } else {
+                        alert(result.data.Error)
+                    }
+                })
+                .catch(err => console.log(err))
+        }
     }
 
     return (
